requirements.txt
pyserial==3.0.1
versioneer==0.16
